luxury-travel-curator/requirements.txt
python-dotenv
google-generativeai
requests
beautifulsoup4
